converter: five silent miscompiles the regexp pair exposed — all compiled, none behaved
Every one of these produced C# that built clean and then did the wrong thing at
run time; each was root-caused against the emitted `.cs` and fixed at its own
layer. Together they take `regexp/syntax` from 6/12 to 12/12 and `regexp` from
12/44 to 45/45.

1. A CONSTANT emitted as `static readonly` is an initialization dependency too
   (initOrderOperations.go, visitValueSpec.go). Go constants are compile-time and
   `types.Info.InitOrder` never mentions them, but a const of a named, string,
   complex, uintptr or untyped type has no C# `const` form — it is a plain
   `static readonly` FIELD, and so carries the same cross-part
   field-initializer ordering hazard a package var does. `regexp/syntax`'s
   `opNames` table, index-keyed by the `Op` constants declared in another file,
   read every key as zero and collapsed 19 entries into slot 0, so `dumpRegexp`
   took its numeric fallback for every operator: `op3{a}` where Go prints
   `lit{a}`. The relocation analysis now carries a const dependency edge
   (transitive, on the same memoized graph), and the emission side grew the
   matching half: relocation lived only in the non-constant initializer arm, yet
   a Go-CONSTANT-valued initializer still renders the source expression and so
   still reads the field. Corpus effect: nine packages gained a
   `package_init.cs`, each a latent zero-value const read.

2. A BLANK result mixed with a named one still needs the named-return-defer
   handling (iifeOperations.go, visitFuncDecl.go, main.go). Go allows
   `(_ *Regexp, err error)`, and deferred code can still mutate `err`. Rejecting
   the signature on the first `_` dropped the whole mechanism, so a recovered
   panic returned the ZERO tuple — `parse` reported (nil, nil), a SUCCESSFUL
   parse of every expression that must fail. The blank slot now gets a generated
   local (C#'s `_` is the discard, so it cannot be declared under that name),
   interned per result object.

3. A write that ENCLOSES the literal counts as written-after-capture
   (variableAnalysisOperations.go). `check = func(...){ ... check(...) ... }` —
   Go's recursive-closure idiom — has its store positioned BEFORE the literal it
   contains, so the position-only scan scored it read-only and took the snapshot
   path, capturing the still-null delegate. Every recursion NREd; `makeOnePass`
   is that function, so `^.$` and most of the package became uncompilable.

4. Assigning `nil` to a pointer PARAMETER is a box repoint, not a write to the
   pointee (visitAssignStmt.go). The untyped nil literal has no pointer type, so
   the box-reassignment triggers missed it and `p = nil` ZEROED the pointed-to
   struct while leaving `Ꮡp` non-nil. `makeOnePass`'s bail-out then handed
   `compileOnePass` an onePassProg with an emptied Inst slice — index-out-of-range
   in `cleanupOnePass` for every pattern the analysis rejected.

5. A SELECTOR left-hand side counts as a reassignment (visitAssignStmt.go). A
   field write is a write to existing storage exactly as the index and star-deref
   forms are — the reference doc already said so — but it was counted as neither
   reassigned nor declared, so a parallel assignment whose targets are all
   selectors matched no tuple-path gate and shattered into sequential stores.
   `inst.Out, inst.Arg = inst.Arg, inst.Out` left BOTH fields holding the
   original Arg (the very next line's plain-local swap was already correct), so
   regexp quietly lost its one-pass engine and `^[a-c]+$` stopped matching.

Guards, all five proven to FAIL on the pre-fix converter while compiling cleanly
there — extensions to the projects that already own each mechanism:
PackageVarInitOrder (named-type const table + constant-folded initializer),
NamedReturnDefer (`(_ *box, err error)` with a recovering defer),
ClosureWriteVisibility Q1/Q2 (self- and mutually-recursive closures),
PointerParamNilWalk (`dropIfShort` nils the parameter, caller proves the pointee
survived), ParallelAssignmentHazard (field swap, cross-struct rotate, package-var
swap).

Gates: check-no-regression classified — the only goldens that move are the five
guard projects plus AndNotAssignNarrow and RangeVarReassign, both consequences of
(5) and both verified Output-identical to `go run`; full behavioral suite
502/502; seeded whole-stdlib reconvert (34 marked files, 0 clobbered, path-precise
content gate with a firing positive control) and a 305-project corpus build with
0 errors; operational re-validation of every banked package.

Docs: ConversionStrategies-Reference gains a subsection per root cause under the
matching topic, and the summary's multi-assignment section gains the field-swap
example.

+++ src/Tests/Behavioral/AndNotAssignNarrow/main.cs
@@ -12,8 +12,8 @@ internal static readonly UntypedInt writing = 4;
 
 internal static void Main() {
     var h = Ꮡ(new hmap(flags: 7));
-    h.Value.flags &= unchecked((uint8)~writing);
-    h.Value.flags |= 8;
+    h.Value.flags &= unchecked((uint8)~(uint8)(writing));
+    h.Value.flags |= (uint8)(8);
     fmt.Println((~h).flags);
     uint16 u = 0xFFFF;
     u &= unchecked((uint16)~(uint16)(0x0F0));
